Cleaned up compiler warnings on Win32 from Borland compiler.

+++ bacnet-stack/dlmstp.h
@@ -49,7 +49,7 @@ typedef struct dlmstp_packet {
     bool ready;                 /* true if ready to be sent or received */
     BACNET_ADDRESS address;     /* source address */
     uint8_t frame_type;         /* type of message */
-    unsigned pdu_len;           /* packet length */
+    uint16_t pdu_len;           /* packet length */
     uint8_t pdu[MAX_MPDU];      /* packet */
 } DLMSTP_PACKET;
